refactor piezo_movement_correction from Analysis.points to Analysis
